Added palette cycling.
If the palette cycle checkbox is checked, the timerPaletteCycle_ticket
method is called every 100 milliseconds. at each call, the current bitmap
is saved in memory as a gif image, to access the original bitmaps Palette
Entries (256 colours, not accessible in jpg). Each palette entry (colour)
is set to the next entry in the array. The colour palette is added to a
new bitmap which is drawn.

Removed unused ticks variable.

Changed name of reset method.