Animate through a sequence of fractals.
User can select multiple bitmap states in xml from the dialog menu, which
are stored in the States list. Each time the timerSequenceOfFractals_Tick
method is called, ticks is incremented. The vaue of ticks is used to
select the next state from the States list.

The user can change the timerSequenceOfFractals timer interval using the
domainUpDown form control.

The user can use the trackbar to switch between states.